+++ gazebo/physics/Gripper.hh
@@ -14,23 +14,20 @@
  * limitations under the License.
  *
 */
-#ifndef _GRIPPER_HH_
-#define _GRIPPER_HH_
+#ifndef GAZEBO_PHYSICS_GRIPPER_HH_
+#define GAZEBO_PHYSICS_GRIPPER_HH_
 
-#include <map>
-#include <vector>
 #include <string>
 
-#include "gazebo/msgs/msgs.hh"
-#include "gazebo/transport/TransportTypes.hh"
-#include "gazebo/math/Pose.hh"
-#include "gazebo/physics/PhysicsTypes.hh"
 #include "gazebo/util/system.hh"
 
 namespace gazebo
 {
   namespace physics
   {
+    // Forward declare private data class
+    class GripperPrivate;
+
     /// \addtogroup gazebo_physics
     /// \{
 
@@ -59,102 +56,21 @@ namespace gazebo
       public: virtual void Init();
 
       /// \brief Return the name of the gripper.
-      public: std::string GetName() const;
+      /// \deprecated See std::string Name() const;
+      public: std::string GetName() const GAZEBO_DEPRECATED(8.0);
+
+      /// \brief Return the name of the gripper.
+      /// \return Name of the gripper
+      public: std::string Name() const;
 
       /// \brief True if the gripper is attached to another model.
       /// \return True if the gripper is active and a joint has been
       /// created between the gripper and another model.
       public: bool IsAttached() const;
 
-      /// \brief Update the gripper.
-      private: void OnUpdate();
-
-      /// \brief Callback used when the gripper contacts an object.
-      /// \param[in] _msg Message that contains contact information.
-      private: void OnContacts(ConstContactsPtr &_msg);
-
-      /// \brief Attach an object to the gripper.
-      private: void HandleAttach();
-
-      /// \brief Detach an object from the gripper.
-      private: void HandleDetach();
-
-      /// \brief A reset function.
-      private: void ResetDiffs();
-
-      /// \brief Model that contains this gripper.
-      private: physics::ModelPtr model;
-
-      /// \brief The physics engine.
-      // NOTE: Variable not used, left here for ABI compatibility. Remove when
-      // merging forward.
-      private: physics::PhysicsEnginePtr physics;
-
-      /// \brief Pointer to the world.
-      private: physics::WorldPtr world;
-
-      /// \brief A fixed joint to connect the gripper to a grasped object.
-      private: physics::JointPtr fixedJoint;
-
-      /// \brief The base link for the gripper.
-      private: physics::LinkPtr palmLink;
-
-      /// \brief All our connections.
-      private: std::vector<event::ConnectionPtr> connections;
-
-      /// \brief The collisions for the links in the gripper.
-      private: std::map<std::string, physics::CollisionPtr> collisions;
-
-      /// \brief The current contacts.
-      private: std::vector<msgs::Contact> contacts;
-
-      /// \brief Mutex used to protect reading/writing the sonar message.
-      private: boost::mutex mutexContacts;
-
-      /// \brief True if the gripper has an object.
-      private: bool attached;
-
-      /// \brief Previous difference between the palm link and grasped
-      /// object.
-      private: math::Pose prevDiff;
-
-      /// \brief Used to determine when to create the fixed joint.
-      private: std::vector<double> diffs;
-
-      /// \brief Current index into the diff array.
-      private: int diffIndex;
-
-      /// \brief Rate at which to update the gripper.
-      private: common::Time updateRate;
-
-      /// \brief Previous time when the gripper was updated.
-      private: common::Time prevUpdateTime;
-
-      /// \brief Number of iterations the gripper was contacting the same
-      /// object.
-      private: int posCount;
-
-      /// \brief Number of iterations the gripper was not contacting the same
-      /// object.
-      private: int zeroCount;
-
-      /// \brief Minimum number of links touching.
-      private: unsigned int minContactCount;
-
-      /// \brief Steps touching before engaging fixed joint
-      private: int attachSteps;
-
-      /// \brief Steps not touching before disengaging fixed joint
-      private: int detachSteps;
-
-      /// \brief Name of the gripper.
-      private: std::string name;
-
-      /// \brief Node for communication.
-      protected: transport::NodePtr node;
-
-      /// \brief Subscription to contact messages from the physics engine.
-      private: transport::SubscriberPtr contactSub;
+      /// \internal
+      /// \brief Private data pointer
+      private: std::unique_ptr<GripperPrivate> dataPtr;
     };
     /// \}
   }